requirements.txt
Flask==2.2.3
bcrypt==4.0.1
SQLAlchemy==2.0.9
python-dotenv==1.0.0
PyJWT~=2.6.0